Move render code to its own module
